Replace 'long int' with 'int' for Alpha. This change should have no effect on i386.
Pointed out by: Steve Kargl <sgk@troutmask.apl.washington.edu>

Quote from http://www.netlib.org/f2c/readme:

NOTE:   f2c.h defines several types, e.g., real, integer, doublereal.
        The definitions in f2c.h are suitable for most machines, but if
        your machine has sizeof(double) > 2*sizeof(long), you may need
        to adjust f2c.h appropriately.  f2c assumes
                sizeof(doublecomplex) = 2*sizeof(doublereal)
                sizeof(doublereal) = sizeof(complex)
                sizeof(doublereal) = 2*sizeof(real)
                sizeof(real) = sizeof(integer)
                sizeof(real) = sizeof(logical)
                sizeof(real) = 2*sizeof(shortint)
        EQUIVALENCEs may not be translated correctly if these
        assumptions are violated.

        On machines, such as those using a DEC Alpha processor, on
        which sizeof(short) == 2, sizeof(int) == sizeof(float) == 4,
        and sizeof(long) == sizeof(double) == 8, it suffices to
        modify f2c.h by removing the first occurrence of "long "
        on each line containing "long ", e.g., by issuing the
        commands
                mv f2c.h f2c.h0
                sed 's/long //' f2c.h0 >f2c.h
        On such machines, one can enable INTEGER*8 by uncommenting
        the typedef of longint in f2c.h, so it reads
                typedef long longint;
        by compiling libI77 with -DAllow_TYQUAD, and by adjusting
        libF77/makefile as described in libF77/README.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@42825 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ usr.bin/f2c/f2c.h
@@ -7,15 +7,15 @@
 #ifndef F2C_INCLUDE
 #define F2C_INCLUDE
 
-typedef long int integer;
-typedef unsigned long uinteger;
+typedef int integer;
+typedef unsigned int uinteger;
 typedef char *address;
 typedef short int shortint;
 typedef float real;
 typedef double doublereal;
 typedef struct { real r, i; } complex;
 typedef struct { doublereal r, i; } doublecomplex;
-typedef long int logical;
+typedef int logical;
 typedef short int shortlogical;
 typedef char logical1;
 typedef char integer1;
@@ -42,9 +42,9 @@ typedef short flag;
 typedef short ftnlen;
 typedef short ftnint;
 #else
-typedef long int flag;
-typedef long int ftnlen;
-typedef long int ftnint;
+typedef int flag;
+typedef int ftnlen;
+typedef int ftnint;
 #endif
 
 /*external read, write*/
